requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cmake==4.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
Django==5.2.9
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
dlib @ https://github.com/z-mahmud22/Dlib_Windows_Python3.x/raw/main/dlib-19.24.99-cp312-cp312-win_amd64.whl#sha256=20c62e606ca4c9961305f7be3d03990380d3e6c17f8d27798996e97a73271862
et_xmlfile==2.0.0
executing==2.2.1
face-recognition==1.3.0
face_recognition_models==0.3.0
fastapi==0.120.3
ffmpy==0.6.4
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
gast==0.6.0
google-pasta==0.2.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
iniconfig==2.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyterlab_widgets==3.0.16
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pbr==6.1.1
pgm==0.0.2
pillow==11.3.0
pluggy==1.6.0
ply==3.11
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
pure_eval==0.2.3
pydantic==2.11.10
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
qiskit==2.2.3
qiskit-aer==0.17.2
requests==2.32.5
rich==14.2.0
ruff==0.14.3
rustworkx==0.16.0
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.0
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.49.1
stevedore==5.4.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==2.0.0
